+++ app.js
@@ -1,9 +1,15 @@
 const express = require("express");
+
 const logger = require("morgan");
+
 const cors = require("cors");
+
 require("dotenv").config();
+
 const contactsRouter = require("./routes/api/contacts");
+
 const app = express();
+
 const formatsLogger = app.get("env") === "development" ? "dev" : "short";
 
 app.use(logger(formatsLogger));
